Crate atspi_common

atspi-common

Defines all common types, events, and data structures for atspi-proxies and atspi-connection. Since atspi-proxies and atspi-connection are downstream crates, the documentation can not link to it directly. Any type ending in *Proxy is in atspi-proxies.

Re-exports

pub use crate::events::event_wrappers::CacheEvents;

pub use crate::events::event_wrappers::DocumentEvents;

pub use crate::events::event_wrappers::Event;

pub use crate::events::event_wrappers::EventListenerEvents;

pub use crate::events::event_wrappers::FocusEvents;

pub use crate::events::event_wrappers::KeyboardEvents;

pub use crate::events::event_wrappers::MouseEvents;

pub use crate::events::event_wrappers::ObjectEvents;

pub use crate::events::event_wrappers::TerminalEvents;

pub use crate::events::event_wrappers::WindowEvents;

pub use action::Action;

pub use object_match::MatchType;

pub use object_match::ObjectMatchRule;

pub use object_match::SortOrder;

pub use object_match::TreeTraversalType;

pub use object_ref::ObjectRef;

pub use object_ref::ObjectRefOwned;

pub use operation::Operation;

pub use interface::Interface;

pub use interface::InterfaceSet;

pub use state::State;

pub use state::StateSet;

pub use cache::CacheItem;

pub use cache::LegacyCacheItem;

pub use error::AtspiError;

pub use events::EventProperties;

pub use events::EventTypeProperties;

Modules

action

cache

Common types for org.a11y.atspi.Cache events.

error

events

interface

Conversion functions and types representing a set of Interfaces.

macros 🔒

object_match

object_ref

operation

relation_type 🔒

role 🔒

state

Structs

TextSelection

Describes a selection of text, including selections across object boundaries.

Enums

ClipType

Enumeration used by TextProxy to indicate how to treat characters intersecting bounding boxes.

CoordType

The coordinate type encodes the frame of reference.

Granularity

Level of granularity to get text of, in relation to a cursor position.

Layer

Indicates relative stacking order of a atspi_proxies::component::ComponentProxy with respect to the onscreen visual representation of the UI.

Politeness

Enumeration used to indicate a type of live region and how assertive it should be in terms of speaking notifications.

RelationType

Describes a relationship between one object and another.

Role

An accessible object role.

ScrollType

Enumeration used by interface the crate::interface::Interface::Accessible to specify where an object should be placed on the screen when using scroll_to.

Type Aliases

Result