atspi_common

Module events

Re-exports

pub use event_body::EventBody;

pub use event_body::EventBodyBorrowed;

pub use event_body::EventBodyOwned;

pub use event_body::EventBodyQtBorrowed;

pub use event_body::EventBodyQtOwned;

pub use event_wrappers::CacheEvents;

pub use event_wrappers::DocumentEvents;

pub use event_wrappers::Event;

pub use event_wrappers::EventListenerEvents;

pub use event_wrappers::FocusEvents;

pub use event_wrappers::KeyboardEvents;

pub use event_wrappers::MouseEvents;

pub use event_wrappers::ObjectEvents;

pub use event_wrappers::TerminalEvents;

pub use event_wrappers::WindowEvents;

pub use traits::DBusInterface;

pub use traits::DBusMatchRule;

pub use traits::DBusMember;

pub use traits::DBusProperties;

pub use traits::EventProperties;

pub use traits::EventTypeProperties;

pub use traits::RegistryEventString;

pub use traits::MessageConversion;

pub use traits::MessageConversionExt;

Modules

cache

document

event_body

event_wrappers

focus

keyboard

mouse

object

All events which can be received by the org.a11y.atspi.Object interface.

registry

This module contains the events that are emitted by the registry daemon. The events are EventListenerRegisteredEvent and EventListenerDeregisteredEvent.

terminal

traits

window