atspi_common::events

Module event_wrappers

Enums

CacheEvents

All events related to the org.a11y.atspi.Cache interface. Note that these are not telling the client that an item has been added to a cache. It is telling the client “here is a bunch of information to store it in your cache”.

DocumentEvents

Event

Encapsulates the various different accessibility bus signal types.

EventListenerEvents

The events that can be emitted by the registry daemon. This enum is used to wrap the events that are emitted by the registry daemon. The events are EventListenerRegisteredEvent and EventListenerDeregisteredEvent.

FocusEvents

KeyboardEvents

MouseEvents

ObjectEvents

TerminalEvents

All events related to the org.a11y.atspi.Event.Terminal interface.

WindowEvents

All events on the org.a11y.atspi.Event.Window interface.