atspi_common::events

Module traits

Traits

DBusInterface

The DBus interface name for an event - or a wrapper type. For example, for any event within the “Object” interface, this should be “org.a11y.atspi.Event.Object”.

DBusMatchRule

A static DBus match rule string. This should usually be a string that looks like this: "type='signal',interface='org.a11y.atspi.Event.Object',member='PropertyChange'";

DBusMember

The DBus member for the event. For example, for an crate::events::object::TextChangedEvent this should be "TextChanged"

DBusProperties

A ‘alias’-trait that combines all the DBus related traits.

EventProperties

EventProperties allows access to the internals of an event, specifically:

EventTypeProperties

Describes properties of a specific event type.

EventWrapperMessageConversion 🔒

A way to convert a zbus::Message without checking its interface.

MessageConversion

MessageConversionExt

RegistryEventString

A static Registry event string for registering with the RegistryProxy for receiving events.

TryFromMessage 🔒