atspi_common::events::traits

Trait EventTypeProperties

pub trait EventTypeProperties {
    // Required methods
    fn member(&self) -> &'static str;
    fn interface(&self) -> &'static str;
    fn match_rule(&self) -> &'static str;
    fn registry_string(&self) -> &'static str;
}

Describes properties of a specific event type.

DBus member name

DBus interface name

Together, the member and interface name can describe a specific event type. Likewise, the path and sender bus name collectively make up an ObjectRef, which is a way to uniquely identify an individual accessible item available to atspi. The latter is available via the EventProperties trait.

This can also be generalized, for example this is implemented for crate::Event by dispatching to the matching variants. NOTE: to use EventProperties on wrapper types, like Event, you must enable the “enum-dispatch” feature.

This trait is object-safe.

Required Methods

fn member(&self) -> &'static str

fn interface(&self) -> &'static str

fn match_rule(&self) -> &'static str

fn registry_string(&self) -> &'static str