atspi_common::events::traits

Trait HasInterfaceName

pub trait HasInterfaceName {
    const DBUS_INTERFACE: &'static str;
}

A specific trait only to define an interface name. This is useful for event wrappers like crate::events::ObjectEvents, which, while it does not have other information required to implement the crate::BusProperties trait, you can indeed attach in interface name for all sub events of crate::events::ObjectEvents.

This trait is not object-safe.

Required Associated Constants

const DBUS_INTERFACE: &'static str

A static interface string for DBus. This should usually be a string that looks like this: "org.a11y.atspi.Event.*";

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.