atspi_common

Module macros

Macros

event_test_cases 🔒

event_wrapper_test_cases 🔒

Expands to five tests:

impl_event_properties 🔒

Expands to implement the required methods for the crate::EventProperties trait. This depends on the struct to have an item field of type crate::ObjectRef.

impl_event_type_properties_for_event 🔒

Implement EventTypeProperties for a given event type.

impl_from_dbus_message 🔒

Implements the TryFrom trait for a given event type. Converts a zbus::Message into a user facing event type.

impl_from_interface_event_enum_for_event 🔒

Expands to a conversion given the enclosed event type and outer Event variant.

impl_from_object_ref 🔒

Expands to implement From for crate::ObjectRef. This depends on the struct to have an item field of type crate::ObjectRef.

impl_from_user_facing_event_for_interface_event_enum 🔒

Expands to a conversion given the user facing event type, the wrapping interface enum variant, and the outer Event variant.

impl_from_user_facing_type_for_event_enum 🔒

Expands to a conversion given two arguments,

impl_member_interface_registry_string_and_match_rule_for_event 🔒

Implement DBusMember, DBusInterface, DBusMatchRule, and RegistryEventString for a given event type.

impl_msg_conversion_ext_for_target_type 🔒

Implements MessageConversionExt for a given target event type.

impl_msg_conversion_ext_for_target_type_with_specified_body_type 🔒

Implements MessageConversionExt for a given target event type with a given body type.

impl_msg_conversion_for_types_built_from_object_ref 🔒

Implement the MessageConversion trait for the given types.

impl_to_dbus_message 🔒

Implements the TryFrom trait for a given event type. Converts a user facing event type into a zbus::Message.

impl_try_from_event_for_interface_enum 🔒

Expands to a conversion given the enclosed event enum type and outer Event variant.

impl_try_from_event_for_user_facing_type 🔒

Expands to a TryFrom<Event> for T where T is the user facing type. The macro takes three arguments:

impl_tryfrommessage_for_event_wrapper 🔒

Implements TryFromMessage for a given event wrapper type.