atspi_common::macros

Macro impl_event_properties

macro_rules! impl_event_properties {
    ($type:ty) => { ... };
}

Expands to implement the required methods for the crate::EventProperties trait. This depends on the struct to have an item field of type crate::ObjectRef.

impl_from_interface_event_enum_for_event!(TextCaretMovedEvent);

Expands to:

impl EventProperties for TextCaretMovedEvent {
    fn sender(&self) -> UniqueName<'_> {
        self.item
            .name()
            .expect("Events are constructed with valid ObjetRef")
            .as_ref()
    }
    fn path(&self) -> ObjectPath<'_> {
        self.item.path().as_ref()
    }
}