atspi_common::object_ref

Struct ObjectRefOwned

pub struct ObjectRefOwned(/* private fields */);

A wrapper around the static variant of ObjectRef. This is guaranteed to have a 'static lifetime.

Implementations

impl ObjectRefOwned

pub const fn new(object_ref: ObjectRef<'static>) -> Self

Create a new ObjectRefOwned from an ObjectRef<'static>.

pub const fn from_static_str_unchecked( name: &'static str, path: &'static str, ) -> Self

Create a new ObjectRefOwned from &'static str unchecked.

Safety

The caller must ensure that the strings are valid.

pub fn is_null(&self) -> bool

Returns true if the object reference is Null, otherwise returns false.

pub fn into_inner(self) -> ObjectRef<'static>

Returns the inner ObjectRef, consuming self.

pub fn name(&self) -> Option<&UniqueName<'static>>

Returns the name of the object reference. If the object reference is Null, it returns None. If the object reference is Owned or Borrowed, it returns the name.

Example

use zbus::names::UniqueName;
use zbus::zvariant::ObjectPath;
use atspi_common::ObjectRef;

let name = UniqueName::from_static_str_unchecked(":1.23");
let path = ObjectPath::from_static_str_unchecked("/org/a11y/example/path/007");
let object_ref = ObjectRef::new_borrowed(name, path);

// Check the name of the object reference
assert!(object_ref.name().is_some());
assert_eq!(object_ref.name_as_str().unwrap(), ":1.23");

pub fn path(&self) -> &ObjectPath<'static>

Returns the path of the object reference.
If the object reference is Null, it returns the null-path.

Example

use zbus::names::UniqueName;
use zbus::zvariant::ObjectPath;
use atspi_common::ObjectRef;

let name = UniqueName::from_static_str_unchecked(":1.23");
let path = ObjectPath::from_static_str_unchecked("/org/a11y/example/path/007");
let object_ref = ObjectRef::new_borrowed(name, path);

assert_eq!(object_ref.path_as_str(), "/org/a11y/example/path/007");

pub fn name_as_str(&self) -> Option<&str>

Returns the name of the object reference as a string slice.

pub fn path_as_str(&self) -> &str

Returns the path of the object reference as a string slice.