Crate atspi_common

atspi-common

Defines all common types, events, and data structures for atspi-proxies and atspi-connection. Since atspi-proxies and atspi-connection are downstream crates, the documentation can not link to it directly. Any type ending in *Proxy is in atspi-proxies.

Re-exports

pub use object_ref::ObjectRef;

pub use interface::Interface;

pub use interface::InterfaceSet;

pub use state::State;

pub use state::StateSet;

pub use cache::CacheItem;

pub use cache::LegacyCacheItem;

pub use error::AtspiError;

pub use events::BusProperties;

pub use events::Event;

pub use events::EventProperties;

pub use events::EventTypeProperties;

Modules

cache

Common types for org.a11y.atspi.Cache events.

error

events

interface

Conversion functions and types representing a set of Interfaces.

object_ref

state

Enums

ClipType

Enumeration used by TextProxy to indicate how to treat characters intersecting bounding boxes.

CoordType

The coordinate type encodes the frame of reference.

Granularity

Level of granularity to get text of, in relation to a cursor position.

Layer

Indicates relative stacking order of a atspi_proxies::component::ComponentProxy with respect to the onscreen visual representation of the UI.

Live

Enumeration used to indicate a type of live region and how assertive it should be in terms of speaking notifications. Currently, this is only used for Announcement events, but it may be used for additional purposes in the future. The argument in the Announcement event is named politeness.

MatchType

Enumeration used by MatchArgs to specify how to interpret ObjectRef objects.

RelationType

Describes a relationship between one object and another.

Role

An accessible object role. To think of it in terms of HTML, any semantic element likely has a corollary in this enum. For example: <button>, <input>, <form> or <h4>. Non-semantic elements like <span>, <div> and <b> will not be represented here, and this information is not passed through via the atspi library. TODO: add examples for GTK/Qt libraries in addition to HTML examples.

ScrollType

Enumeration used by interface the crate::interface::Interface::Accessible to specify where an object should be placed on the screen when using scroll_to.

SortOrder

Enumeration used by interface CollectionProxy to specify the way ObjectRef objects should be sorted.

TreeTraversalType

Method of traversing a tree in the CollectionProxy.

Type Aliases

MatchArgs

Result