atspi_common

Module events

Re-exports

pub use event_body::EventBodyBorrowed;

pub use event_body::EventBodyOwned;

pub use event_body::EventBodyQtBorrowed;

pub use event_body::EventBodyQtOwned;

pub use event_wrappers::CacheEvents;

pub use event_wrappers::DocumentEvents;

pub use event_wrappers::Event;

pub use event_wrappers::FocusEvents;

pub use event_wrappers::KeyboardEvents;

pub use event_wrappers::MouseEvents;

pub use event_wrappers::ObjectEvents;

pub use event_wrappers::TerminalEvents;

pub use event_wrappers::WindowEvents;

pub use traits::*;

Modules

cache

document

event_body

event_wrappers

focus

keyboard

mouse

object

terminal

traits

window

Structs

AvailableEvent

An event that is emitted when the registry daemon has started.

EventListenerDeregisteredEvent

An event that is emitted by the registry daemon, to inform that an event has been deregistered to no longer listen for.

EventListenerRegisteredEvent

An event that is emitted by the regostry daemon to signal that an event has been registered to listen for.

EventListeners

Signal type emitted by EventListenerRegistered and EventListenerDeregistered signals, which belong to the Registry interface, implemented by the registry-daemon.

Enums

EventListenerEvents

Covers both EventListener events.