atspi_common

Module events

Modules

document

focus

keyboard

mouse

object

terminal

window

Structs

AddAccessibleEvent

Type that contains the zbus::Message for meta information and the crate::cache::CacheItem

AvailableEvent

An event that is emitted when the registry daemon has started.

EventBody

A borrowed body for events.

EventBodyOwned

Standard event body (GTK, egui, etc.) NOTE: Qt has its own signature: EventBodyQT. Signature (siiva{sv}),

EventBodyQT

Qt event body, which is not the same as other GUI frameworks. Signature: “siiv(so)”

EventListenerDeregisteredEvent

An event that is emitted by the registry daemon, to inform that an event has been deregistered to no longer listen for.

EventListenerRegisteredEvent

An event that is emitted by the regostry daemon to signal that an event has been registered to listen for.

EventListeners

Signal type emitted by EventListenerRegistered and EventListenerDeregistered signals, which belong to the Registry interface, implemented by the registry-daemon.

LegacyAddAccessibleEvent

Type that contains the zbus::Message for meta information and the crate::cache::LegacyCacheItem

RemoveAccessibleEvent

Cache::RemoveAccessible signal event type.

Enums

CacheEvents

All events related to the org.a11y.atspi.Cache interface. Note that these are not telling the client that an item has been added to a cache. It is telling the client “here is a bunch of information to store it in your cache”.

Event

Encapsulates the various different accessibility bus signal types.

EventListenerEvents

Covers both EventListener events.

Constants

ATSPI_EVENT_SIGNATURE

CACHE_ADD_SIGNATURE

EVENT_LISTENER_SIGNATURE

QSPI_EVENT_SIGNATURE

Traits

BusProperties

Describes the DBus-related information about a given struct.

EventProperties

EventProperties allows access to the internals of an event, specifically:

EventTypeProperties

Describes properties of a specific event type.

HasMatchRule

A specific trait only to define match rules. This is useful for event wrappers like ObjectEvents, which, while it does not have other information required to implement the BusProperties trait, you can indeed add a match rule to the DBus connection to capture all sub events of ObjectEvents.

HasRegistryEventString

A specific trait only to define registry event matches. This is useful for event wrappers like ObjectEvents, which, while it does not have other information required to implement the BusProperties trait, you can indeed add a match rule to the AT-SPI connection to subscribe to all sub events of ObjectEvents.