Trait atspi_common::events::HasMatchRule

pub trait HasMatchRule {
    const MATCH_RULE_STRING: &'static str;
}

A specific trait only to define match rules.

Required Associated Constants

const MATCH_RULE_STRING: &'static str

A static match rule string for DBus. This should usually be a string that looks like this: "type='signal',interface='org.a11y.atspi.Event.Object',member='PropertyChange'"; This should be deprecated in favour of composing the string from GenericEvent::DBUS_MEMBER and GenericEvent::DBUS_INTERFACE.

Object Safety

This trait is not object safe.