atspi_common::events

Trait HasMatchRule

pub trait HasMatchRule {
    const MATCH_RULE_STRING: &'static str;
}

A specific trait only to define match rules. This is useful for event wrappers like ObjectEvents, which, while it does not have other information required to implement the BusProperties trait, you can indeed add a match rule to the DBus connection to capture all sub events of ObjectEvents.

This trait is not object-safe.

Required Associated Constants

const MATCH_RULE_STRING: &'static str

A static match rule string for DBus. This should usually be a string that looks like this: "type='signal',interface='org.a11y.atspi.Event.Object',member='PropertyChange'"; This should be deprecated in favour of composing the string from BusProperties::DBUS_MEMBER and BusProperties::DBUS_INTERFACE.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.